✨ Introduce session-close route run background
