+++ app/utils/auto_task_utils.py
@@ -13,3 +13,19 @@ def get_current_step_message(tool: str, status: str) -> str:
     tool_msg = tool_kor.get(tool, tool)
     status_msg = status_kor.get(status, status)
     return f"{tool_msg} {status_msg}"
+
+# Swagger None UI Error 대처
+def safe_background_response(data: dict) -> dict:
+    result = {
+        'user_id': data.get('user_id') or "",
+        'finished': bool(data.get('finished'))
+    }
+    if data.get('task') is not None:
+        result['task'] = data['task']
+    if data.get('last_completed_title'):
+        result['last_completed_title'] = data['last_completed_title']
+    if data.get('error'):
+        result['error'] = data['error']
+    if data.get('step') is not None:
+        result['step'] = data['step']
+    return result
